most of git diff main..tgb-932-can-helper-index  -- tests/test_helpers.py libs/dataclass_utils.py |patch -p1
