+++ tests/test_helpers.py
@@ -20,6 +20,7 @@ from covidactnow.datapublic.common_fields import CommonFields
 from covidactnow.datapublic.common_fields import FieldName
 from covidactnow.datapublic.common_fields import PdFields
 
+from libs.dataclass_utils import dataclass_with_default_init
 from libs.datasets import taglib
 from libs.datasets import timeseries
 from libs.datasets.taglib import TagField
@@ -47,36 +48,6 @@ def _to_list(list_or_scalar: Union[None, T, List[T]]) -> List[T]:
         return [list_or_scalar]
     else:
         return []
-
-
-# TODO(tom): Remove dataclass_with_default_init once we are using Python 3.9. See
-#  https://stackoverflow.com/a/58336722
-def dataclass_with_default_init(_cls=None, *args, **kwargs):
-    def wrap(cls):
-        # Save the current __init__ and remove it so dataclass will
-        # create the default __init__.
-        user_init = getattr(cls, "__init__")
-        delattr(cls, "__init__")
-
-        # let dataclass process our class.
-        result = dataclasses.dataclass(cls, *args, **kwargs)
-
-        # Restore the user's __init__ save the default init to __default_init__.
-        setattr(result, "__default_init__", result.__init__)
-        setattr(result, "__init__", user_init)
-
-        # Just in case that dataclass will return a new instance,
-        # (currently, does not happen), restore cls's __init__.
-        if result is not cls:
-            setattr(cls, "__init__", user_init)
-
-        return result
-
-    # Support both dataclass_with_default_init() and dataclass_with_default_init
-    if _cls is None:
-        return wrap
-    else:
-        return wrap(_cls)
 
 
 @dataclass_with_default_init(frozen=True)
